refactor: 백엔드 totalFaqs 관련하여 코드 개선

+++ backend/sin/src/main/java/sin/sin/controller/FreqQuestionController.java
@@ -25,6 +25,6 @@ public class FreqQuestionController {
 
     @GetMapping("/total")
     public ResponseEntity<Integer> findTotal(){
-        return ResponseEntity.ok().body((int) freqQuestionService.findTotal());
+        return ResponseEntity.ok().body(freqQuestionService.findTotal());
     }
 }
